third_party/WebKit/LayoutTests/http/tests/devtools/indexeddb/live-update-indexeddb-list-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that the IndexedDB database list live updates.

Dumping IndexedDB tree:
    (empty)
Dumping IndexedDB tree:
    database: database1 - http://127.0.0.1:8000
        (no object stores)
Dumping IndexedDB tree:
    database: database1 - http://127.0.0.1:8000
        (no object stores)
    database: database2 - http://127.0.0.1:8000
        (no object stores)
Dumping IndexedDB tree:
    database: database1 - http://127.0.0.1:8000
        Object store: objectStore1
            Index: index1
    database: database2 - http://127.0.0.1:8000
        (no object stores)
Dumping IndexedDB tree:
    database: database1 - http://127.0.0.1:8000
        Object store: objectStore1
            Index: index1
        Object store: objectStore2
            Index: index2
    database: database2 - http://127.0.0.1:8000
        (no object stores)
Dumping IndexedDB tree:
    database: database1 - http://127.0.0.1:8000
        Object store: objectStore1
            Index: index1
            Index: index3
        Object store: objectStore2
            Index: index2
    database: database2 - http://127.0.0.1:8000
        (no object stores)
Dumping IndexedDB tree:
    database: database1 - http://127.0.0.1:8000
        Object store: objectStore1
            Index: index1
        Object store: objectStore2
            Index: index2
    database: database2 - http://127.0.0.1:8000
        (no object stores)
Dumping IndexedDB tree:
    database: database1 - http://127.0.0.1:8000
        Object store: objectStore1
            Index: index1
    database: database2 - http://127.0.0.1:8000
        (no object stores)
Dumping IndexedDB tree:
    database: database2 - http://127.0.0.1:8000
        (no object stores)
Dumping IndexedDB tree:
    (empty)
